Update README.md
Added an important and badly needed philosophical essay in this README.md Markdown file, to this "Interdisciplinary Art Research" Repository on GitHub, called "Interdisciplinarity and the Interwoven Fabric of Artistic Practice and Research", which I just finished composing, in the middle of the night. Yes, the interdisciplinary artist-research never sleeps. :)