Update README.md
Changed the hyperlink in my AG Signature to the historiotheque.ca site instead of my usual alexgagnon.com site, because it's more pertinent to the work I'm doing as an artist, as of now. If I expand the alexgagnon.com website, I might change back to that hyperlink in the future, if I deem it necessary.